+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Next Release
+
+## Bugfixes
+
+* Fix `#connection_completed` not be called after `#connect`.
+
 # v0.1.0
 
 ## Features