+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Next Release
+# v0.1.1
 
 ## Bugfixes